ssd-728 adding func to add address to manufacturer. works with country code instead of id. also refactored into resources directory
