+++ daaily/lucy/constants.py
@@ -1,2 +1,228 @@
 LUCY_V2_BASE_URL_STAGING = "https://lucy.staging.daaily.com/api/v2"
 LUCY_V2_BASE_URL_PRODUCTION = "https://lucy.daaily.com/api/v2"
+
+
+COUNTRY_CODE_TO_COUNTRY_ID_MAPPING = {
+    None: 10000016,
+    "AF": 8992000,
+    "AL": 8992001,
+    "DZ": 8992002,
+    "AS": 8992003,
+    "AD": 8992004,
+    "AO": 8992005,
+    "AG": 8992006,
+    "AR": 8992007,
+    "AM": 8992008,
+    "AW": 8992009,
+    "AU": 8992010,
+    "AT": 8992011,
+    "AZ": 8992012,
+    "BS": 8992013,
+    "BH": 8992014,
+    "BD": 8992015,
+    "BB": 8992016,
+    "BY": 8992017,
+    "BE": 8992018,
+    "BZ": 8992019,
+    "BJ": 8992020,
+    "BM": 8992021,
+    "BT": 8992022,
+    "BA": 8992024,
+    "BW": 8992025,
+    "BR": 8992026,
+    "IO": 8992027,
+    "BN": 8992028,
+    "BG": 8992029,
+    "BF": 8992030,
+    "BI": 8992031,
+    "KH": 8992032,
+    "CM": 8992033,
+    "CA": 8992034,
+    "KY": 8992036,
+    "CF": 8992037,
+    "TD": 8992038,
+    "CL": 8992039,
+    "CN": 8992040,
+    "CO": 8992041,
+    "KM": 8992042,
+    "CK": 8992044,
+    "CR": 8992045,
+    "CI": 8992046,
+    "HR": 8992047,
+    "CU": 8992048,
+    "CY": 8992049,
+    "DK": 8992051,
+    "DJ": 8992052,
+    "DM": 8992053,
+    "DO": 8992054,
+    "EC": 8992056,
+    "EG": 8992057,
+    "SV": 8992058,
+    "GQ": 8992059,
+    "ER": 8992060,
+    "EE": 8992061,
+    "ET": 8992062,
+    "FO": 8992064,
+    "FM": 8992065,
+    "FJ": 8992066,
+    "FI": 8992067,
+    "FR": 8992068,
+    "PF": 8992069,
+    "TF": 8992070,
+    "GA": 8992071,
+    "GM": 8992072,
+    "GE": 8992073,
+    "DE": 8992074,
+    "GH": 8992075,
+    "GI": 8992076,
+    "GR": 8992077,
+    "GL": 8992078,
+    "GD": 8992079,
+    "GP": 8992080,
+    "GU": 8992081,
+    "GT": 8992082,
+    "GN": 8992083,
+    "GW": 8992084,
+    "GY": 8992085,
+    "HT": 8992086,
+    "HN": 8992088,
+    "HK": 8992089,
+    "HU": 8992090,
+    "IS": 8992091,
+    "IN": 8992092,
+    "ID": 8992093,
+    "IQ": 8992094,
+    "IE": 8992095,
+    "IL": 8992097,
+    "IT": 8992098,
+    "JM": 8992099,
+    "JP": 8992100,
+    "JO": 8992101,
+    "KZ": 8992102,
+    "KE": 8992103,
+    "KI": 8992104,
+    "KW": 8992105,
+    "KG": 8992106,
+    "LA": 8992107,
+    "LV": 8992108,
+    "LB": 8992109,
+    "LS": 8992110,
+    "LR": 8992111,
+    "LY": 8992112,
+    "LI": 8992113,
+    "LT": 8992114,
+    "LU": 8992115,
+    "MO": 8992116,
+    "MG": 8992117,
+    "MW": 8992118,
+    "MY": 8992119,
+    "MV": 8992120,
+    "ML": 8992121,
+    "MT": 8992122,
+    "MH": 8992123,
+    "MQ": 8992124,
+    "MR": 8992125,
+    "MU": 8992126,
+    "MX": 8992127,
+    "MC": 8992128,
+    "MN": 8992129,
+    "MA": 8992130,
+    "MZ": 8992131,
+    "MM": 8992132,
+    "NA": 8992133,
+    "NR": 8992134,
+    "NP": 8992135,
+    "NL": 8992136,
+    "NC": 8992138,
+    "NZ": 8992139,
+    "NI": 8992140,
+    "NE": 8992141,
+    "NG": 8992142,
+    "MP": 8992143,
+    "NO": 8992144,
+    "OM": 8992145,
+    "PK": 8992146,
+    "PW": 8992147,
+    "PA": 8992149,
+    "PG": 8992150,
+    "PY": 8992151,
+    "PE": 8992152,
+    "PH": 8992153,
+    "PL": 8992154,
+    "PT": 8992155,
+    "PR": 8992156,
+    "QA": 8992157,
+    "RE": 8992160,
+    "RO": 8992161,
+    "RU": 8992162,
+    "RW": 8992163,
+    "KN": 8992164,
+    "LC": 8992165,
+    "VC": 8992166,
+    "WS": 8992167,
+    "SM": 8992168,
+    "SA": 8992170,
+    "SN": 8992171,
+    "SC": 8992173,
+    "SL": 8992174,
+    "SG": 8992175,
+    "SK": 8992176,
+    "SI": 8992177,
+    "SB": 8992178,
+    "SO": 8992179,
+    "ZA": 8992180,
+    "ES": 8992181,
+    "LK": 8992182,
+    "SD": 8992183,
+    "SR": 8992184,
+    "SZ": 8992185,
+    "SE": 8992186,
+    "CH": 8992187,
+    "SY": 8992188,
+    "TJ": 8992190,
+    "TH": 8992191,
+    "TG": 8992194,
+    "TK": 8992195,
+    "TO": 8992196,
+    "TT": 8992197,
+    "TN": 8992198,
+    "TR": 8992199,
+    "TM": 8992200,
+    "TC": 8992201,
+    "TV": 8992202,
+    "UG": 8992203,
+    "UA": 8992204,
+    "AE": 8992205,
+    "GB": 8992206,
+    "US": 8992208,
+    "UY": 8992209,
+    "UZ": 8992210,
+    "VU": 8992211,
+    "VG": 8992214,
+    "VI": 8992215,
+    "EH": 8992216,
+    "YE": 8992217,
+    "ZM": 8992218,
+    "ZW": 8992219,
+    "GG": 8992220,
+    "NF": 8992221,
+    "JE": 8992222,
+    "IM": 8992223,
+    "MS": 8992224,
+    "UM": 8992226,
+    "AQ": 8992227,
+    "AI": 8992228,
+    "GF": 8992229,
+    "YT": 8992230,
+    "ME": 8992231,
+    "RS": 8992232,
+    "PM": 8992233,
+    "BV": 8992234,
+    "WF": 8992235,
+    "NU": 8992236,
+    "AX": 8992237,
+    "CW": 8992239,
+    "SS": 8992244,
+    "CX": 8992245,
+    "SJ": 8992246,
+}
