+++ daaily/lucy/resources/manufacturer.py
@@ -1,0 +1,284 @@
+import json
+import mimetypes
+from typing import Any, Dict, Generator, Literal
+
+import urllib3
+
+from daaily.lucy.constants import COUNTRY_CODE_TO_COUNTRY_ID_MAPPING
+from daaily.lucy.enums import EntityType
+from daaily.lucy.models import Filter
+from daaily.lucy.response import Response
+from daaily.lucy.utils import (
+    add_image_to_manufacturer,
+    check_field_content_set,
+    gen_new_image_object_with_extras,
+)
+
+from . import BaseResource
+
+MANUFACTURER_IMAGE_UPLOAD_ENDPOINT = "/manufacturers/{manufacturer_id}/image/upload"
+
+http = urllib3.PoolManager()  # for handling HTTP requests without auth
+
+
+class ManufacturersResource(BaseResource):
+    def get(
+        self, filters: list[Filter] | None = None
+    ) -> Generator[Dict[str, Any], None, None]:
+        """
+        Retrieves manufacturers with optional filtering, returning them as a generator
+        that yields each manufacturer one at a time.
+
+        Available filters:
+            - manufacturer_ids (str): Filter by comma separated manufacturer IDs.
+            - manufacturer_name (str): Filter by manufacturer name.
+
+        Note that the following filters are automatically added to the query:
+            - skip (int): Number of records to skip.
+            - limit (int): Maximum number of records to retrieve.
+
+        Args:
+            filters (list[Filter] | None): A list of filters to apply to the query.
+
+        Yields:
+            dict: A dictionary representing a single manufacturer.
+
+        Example:
+            ```python
+            # Define filters
+            filters = [Filter("manufacturer_ids", "3100099,3100100,3100101")]
+
+            # Get manufacturers (pagination handled internally)
+            manufacturers = client.manufacturers.get(filters=filters)
+
+            # Iterate over the results without worrying about pagination
+            for m in manufacturers:
+                print(f"ID: {m['manufacturer_id']}, Name: {m['name']}")
+            ```
+        """
+        if filters is None:
+            filters = []
+        filters = [f for f in filters if f.name not in ["limit", "skip"]]
+        limit_filter = Filter(name="limit", value="100")
+        skip_filter = Filter(name="skip", value="0")
+        filters.append(limit_filter)
+        filters.append(skip_filter)
+        while True:
+            response = self._client.get_entities(EntityType.MANUFACTURER, filters)
+            if response.status != 200:
+                break
+            for item in response.json():  # type: ignore
+                yield item
+            skip = int(skip_filter.value) + int(limit_filter.value)
+            skip_filter.value = str(skip)
+            filters = [f for f in filters if f.name != "skip"]
+            filters.append(skip_filter)
+
+    def get_by_id(self, manufacturer_id: int):
+        return self._client.get_entity(EntityType.MANUFACTURER, manufacturer_id)
+
+    def get_by_domain(self, domain: str):
+        return self._client.get_entity_custom(EntityType.MANUFACTURER, domain, "domain")
+
+    def update(self, manufacturers: list[dict], filters: list[Filter] | None = None):
+        return self._client.update_entities(
+            EntityType.MANUFACTURER, manufacturers, filters
+        )
+
+    def create(self, manufacturers: list[dict], filters: list[Filter] | None = None):
+        return self._client.create_entities(
+            EntityType.MANUFACTURER, manufacturers, filters
+        )
+
+    def add_address(
+        self,
+        manufacturer_id: int,
+        address: dict,
+        country_code: str | None,
+    ):
+        """
+        Adds an address to a manufacturer.
+
+        This method adds a new address to the manufacturer's information. The address
+        is expected to be a dictionary with several key-value pairs, which define the
+        address details. The country code is used to map the address to the correct
+        country ID.
+
+        Args:
+            manufacturer_id (int): The ID of the manufacturer to which the address will
+                be added.
+            address (dict): The address to be added. The dictionary should include the
+                following keys:
+                - address_type (str|ENUM): The address type, e.g., "headquarter";
+                    "country_representation"; "location".
+                - status (str|ENUM): The status of the address, e.g., "online" or
+                    "deleted".
+                - street (str | None): The street address.
+                - city (str | None): The city of the address.
+                - zip (str | None): The postal code.
+                - zip_box (str | None): The postal box code (if any).
+                - phone (str | None): The phone number associated with the address.
+                - homepage (str | None): The homepage URL associated with the address.
+                - distribution_partner (str | None): The distribution partner's website
+                    (if applicable).
+                - name (str | None): The name of the contact person.
+                - contact_email (str): The contact person's email address.
+                - contact_language (str | None): The language preference of the contact
+                    person (e.g., "de").
+            country_code (str | None): The country code that maps to a valid country ID.
+
+        Raises:
+            ValueError: If the provided country code is invalid or not found in the
+                country code mapping.
+            Exception: If the manufacturer cannot be retrieved or updated.
+
+        Returns:
+            None: This method modifies the manufacturer's address in place.
+
+        Example:
+            ```python
+            # Define manufacturer ID and address
+            man_id = 12345
+            address = {
+                "address_type": "headquarter",
+                "status": "online",
+                "street": "123 Main St",
+                "city": "Berlin",
+                "zip": "10115",
+                "zip_box": "PO Box 123",
+                "phone": "+49 123 456789",
+                "homepage": "http://manufacturer.com",
+                "distribution_partner": "http://distributor.com",
+                "name": "John Doe",
+                "contact_email": "john.doe@manufacturer.com",
+                "contact_language": "de",
+            }
+
+            # Add address to manufacturer
+            client.manufacturers.add_address(man_id, address, "DE")
+            ```
+        """
+        country_id = COUNTRY_CODE_TO_COUNTRY_ID_MAPPING.get(country_code)
+        if country_id is None:
+            raise ValueError(
+                f"Invalid country code: {country_code}. "
+                + f"Must be one of: {COUNTRY_CODE_TO_COUNTRY_ID_MAPPING.keys()}"
+            )
+        address["country_of_user_id"] = country_id
+        address["country_id"] = country_id
+        response = self._client.get_entity(EntityType.MANUFACTURER, manufacturer_id)
+        if response.status != 200:
+            raise Exception(f"Failed to get manufacturer: {response.data}")
+        m = response.json()
+        if not m:
+            raise Exception(f"Manufacturer with ID {manufacturer_id} not found.")
+        if "addresses" not in m:
+            m["addresses"] = []
+        elif address["address_type"] == "headquarter":
+            m["addresses"] = [
+                a for a in m["addresses"] if a["address_type"] != "headquarter"
+            ]
+        m["addresses"].append(address)
+        return self._client.update_entities(EntityType.MANUFACTURER, [m])
+
+    def add_new_image_by_path(
+        self,
+        manufacturer_id: int,
+        image_path: str,
+        image_type: Literal["logo", "header"],
+        image_description: str | None = None,
+        **kwargs,
+    ) -> Response:
+        """
+        Adds an image to a manufacturer by uploading the image from a specified file
+        path.
+
+        This method reads the image file, determines its MIME type, and uploads it to a
+        lucy and then afterwards adds it to the image object on the manufacturer.
+
+        Args:
+            manufacturer_id (int): The ID of the manufacturer to which the image will
+                be added.
+            image_path (str): The file path of the image to be uploaded.
+            image_type (str): The type of image to be uploaded (e.g., "logo", "header").
+            image_description (str | None): Optional description of the image.
+            kwargs (dict | None): Optional fields and metadata add to the image.
+
+        Raises:
+            Exception: If the content type of the image cannot be determined or if the
+            signed URL request fails.
+            Exception: If the manufacturer cannot be retrieved or updated.
+
+        Returns:
+            dict: The updated manufacturer information after adding the image.
+
+        Example:
+            ```python
+            # Define manufacturer ID and image path
+            man_id = 12345
+            image_path = "/path/to/image.jpg"
+
+            # Add image to manufacturer
+            man = client.manufacturers.add_new_image_by_path(man_id, image_path, "logo")
+
+            # Print updated manufacturer information
+            print(man)
+            ```
+        """
+        try:
+            with open(image_path, "rb") as image_file:
+                image_data = image_file.read()
+        except (IOError, OSError) as e:
+            raise Exception(f"Failed to open image file at {image_path}: {e}") from e
+        content_type, _ = mimetypes.guess_type(image_path)
+        if content_type is None:
+            raise Exception(f"Could not determine content type for {image_path}")
+        if not content_type.startswith("image/"):
+            raise Exception(
+                f"File at {image_path} is not an image. Detected: {content_type}"
+            )
+        manufacturer = self._client.get_entity(EntityType.MANUFACTURER, manufacturer_id)
+        if manufacturer.status != 200:
+            raise Exception(f"Failed to get manufacturer: {manufacturer.data}")
+        old_image = check_field_content_set(
+            manufacturer.json(),  # type: ignore
+            f"{image_type}_image",
+        )
+        if kwargs:
+            headers = dict(kwargs.items())
+        else:
+            headers = {}
+        man_image_upload_url = MANUFACTURER_IMAGE_UPLOAD_ENDPOINT.format(
+            manufacturer_id=manufacturer_id
+        )
+        url = f"{self._client._base_url}{man_image_upload_url}?image_type={image_type}"
+        if old_image:
+            url += f"&old_blob_id={old_image['blob_id']}"
+        resp = self._client._do_request(
+            "POST",
+            url,
+            fields={
+                "file": (image_path.split("/")[-1:][0], image_data, content_type),
+            },
+            headers=headers,
+        )
+        if resp.status != 200:
+            raise Exception(
+                f"Failed to upload image. Status code: {resp.status}. {resp.data}"
+            )
+        resp_data = json.loads(resp.data.decode("utf-8"))
+        new_image = gen_new_image_object_with_extras(
+            resp_data["image_blob_id"],
+            size=resp_data["image_size"],
+            height=resp_data["image_height"],
+            width=resp_data["image_width"],
+            file_type=resp_data["image_mime_type"],
+            description=image_description,
+            **kwargs,
+        )
+        manufacturer = add_image_to_manufacturer(
+            manufacturer.json(),  # type: ignore
+            new_image,
+            image_type,
+        )
+        return self._client.update_entities(EntityType.MANUFACTURER, [manufacturer])
