+++ daaily/lucy/resources/project.py
@@ -1,0 +1,68 @@
+from typing import Any, Dict, Generator
+
+from daaily.lucy.enums import EntityType
+from daaily.lucy.models import Filter
+
+from . import BaseResource
+
+
+class ProjectsResource(BaseResource):
+    def get(
+        self, filters: list[Filter] | None = None
+    ) -> Generator[Dict[str, Any], None, None]:
+        """
+        Retrieves project with optional filtering, returning them as a generator
+        that yields each project one at a time.
+
+        Available filters:
+            - project_ids (str): Filter by comma separated project IDs.
+
+        Note that the following filters are automatically added to the query:
+            - skip (int): Number of records to skip.
+            - limit (int): Maximum number of records to retrieve.
+
+        Args:
+            filters (list[Filter] | None): A list of filters to apply to the query.
+
+        Yields:
+            dict: A dictionary representing a single project.
+
+        Example:
+            ```python
+            # Define filters
+            filters = [Filter("manufacturer_id", "12345")]
+
+            # Get materials (pagination handled internally)
+            projects = client.projects.get(filters=filters)
+
+            # Iterate over the results without worrying about pagination
+            for p in projects:
+                print(f"ID: {p['project_id']}, Name: {p['name']}")
+            ```
+        """
+        if filters is None:
+            filters = []
+        filters = [f for f in filters if f.name not in ["limit", "skip"]]
+        limit_filter = Filter(name="limit", value="100")
+        skip_filter = Filter(name="skip", value="0")
+        filters.append(limit_filter)
+        filters.append(skip_filter)
+        while True:
+            response = self._client.get_entities(EntityType.PROJECT, filters)
+            if response.status != 200:
+                break
+            for item in response.json():  # type: ignore
+                yield item
+            skip = int(skip_filter.value) + int(limit_filter.value)
+            skip_filter.value = str(skip)
+            filters = [f for f in filters if f.name != "skip"]
+            filters.append(skip_filter)
+
+    def get_by_id(self, project_id: int):
+        return self._client.get_entity(EntityType.PROJECT, project_id)
+
+    def update(self, projects: list[dict], filters: list[Filter] | None = None):
+        return self._client.update_entities(EntityType.PROJECT, projects, filters)
+
+    def create(self, projects: list[dict], filters: list[Filter] | None = None):
+        return self._client.create_entities(EntityType.PROJECT, projects, filters)
